changed facils booking landing page, reordered directories
